Move GatherCoalescePass to separate file
